Details

Details

Index

ODE problem creation

RigidBodySim.Core.Dynamics — Type.

Dynamics(mechanism)
Dynamics(mechanism, control!; setparams!)

Create a Dynamics object, representing either the passive or closed-loop dynamics of a RigidBodyDynamics.Mechanism.

The control! argument is a callable with the signature control!(τ, t, state), where τ is the torque vector to be set in the body of control!, t is the current time, and state is a MechanismState object. By default, control! is zero_control! (resulting in the passive dynamics).

The setparams! keyword argument is a callable with the signature setparams!(state, p) where state is a MechanismState and p is a vector of parameters, as used in OrdinaryDiffEq.jl.

source

DiffEqBase.ODEProblem — Method.

ODEProblem(dynamics, x0, tspan)
ODEProblem(dynamics, x0, tspan, p; callback, kwargs...)

Create a DiffEqBase.ODEProblem associated with the dynamics of a RigidBodyDynamics.Mechanism.

The initial state x0 can be either a RigidBodyDynamics.MechanismState), or an AbstractVector containing the initial state represented as [q; v; s], where q is the configuration vector, v is the velocity vector, and s is the vector of additional states.

The callback keyword argument can be used to pass in additional DifferentialEquations.jl callbacks.

source

Control

RigidBodySim.Core.zero_control! — Function.

zero_control!(τ, t, state)

A 'zero' controller, i.e. one that sets all control torques to zero at all times.

source

RigidBodySim.Core.controlcallback — Function.

controlcallback(control!)

Can be used to create a callback associated with a given controller.

source

RigidBodySim.Control.PeriodicController — Type.

struct PeriodicController{Tau<:(AbstractArray{T,1} where T), T<:Number, C, I}

A PeriodicController can be used to simulate a digital controller that runs at a fixed rate (in terms of simulation time). It does so by performing a zero-order hold on a provided control function.

PeriodicControllers can be constructed using

PeriodicController(τ, Δt, control!; initialize = DiffEqBase.INITIALIZE_DEFAULT, save_positions = (false, false))

where control! is a controller satisfying the standard RigidBodySim controller signature (control!(τ, Δt, state)), Δt is the simulation time interval between calls to the control! function, and τ is used to call control!. The initialize and save_positions keyword arguments are documented in the DiscreteCallback section of the DifferentialEquations documentation.

PeriodicControllers are callable objects, and themselves fit the standard RigidBodySim controller signature.

A DiffEqCallbacks.PeriodicCallback can be created from a PeriodicController, and is used to stop ODE integration exactly every Δt seconds, so that the control! function can be called. Typically, users will not have to explicitly create this PeriodicCallback, as it is automatically created and added to the ODEProblem when the PeriodicController is passed into the RigidBodySim-provided DiffEqBase.ODEProblem constructor overload.

Examples

In the following example, a PeriodicController is used to simulate a digital PD controller running at a fixed rate of 200 Hz.

julia> using RigidBodySim, RigidBodyDynamics, OrdinaryDiffEq

julia> mechanism = parse_urdf(Float64, Pkg.dir("RigidBodySim", "test", "urdf", "Acrobot.urdf"));

julia> state = MechanismState(mechanism);

julia> set_configuration!(state, [0.1; 0.2]);

julia> controlcalls = Ref(0);

julia> pdcontrol!(τ, t, state) = (controlcalls[] += 1; τ .= -20 .* velocity(state) .- 100 .* configuration(state));

julia> τ = zeros(velocity(state)); Δt = 1 / 200
0.005

julia> problem = ODEProblem(Dynamics(mechanism, PeriodicController(τ, Δt, pdcontrol!)), state, (0., 5.));

julia> sol = solve(problem, Tsit5());

julia> sol.u[end]
4-element Array{Float64,1}:
 -3.25923e-5
 -1.67942e-5
  8.16715e-7
  1.55292e-8

julia> @assert all(x -> isapprox(x, 0, atol = 1e-4), sol.u[end]) # ensure state converges to zero

julia> controlcalls[]
1001

source

Visualization

User interface

DiffEqBase.CallbackSet — Method.

CallbackSet(vis, state; max_fps)

Create the DifferentialEquations.jl callbacks needed for publishing to and receiving commands from a visualizer during simulation.

max_fps is the maximum number of frames per second (in terms of wall time) to draw. Default: 60.0.

source

RigidBodySim.Visualization.VisualizerInterface.visualize — Function.

visualize(vis, t::Number, state::MechanismState)

Visualize a Mechanism at the given time and in the given state using visualizer vis.

source

RigidBodySim.Visualization.VisualizerInterface.window — Function.

window(vis; kwargs...)

Create a new window for visualizer vis. Visualizer-specific keyword arguments (kwargs) may be passed in.

source

RigidBodySim.Visualization.animate — Function.

animate(vis, state, sol; max_fps, realtime_rate, pause_pollint)

Play back a visualization of a RigidBodySim.jl simulation.

Positional arguments:

vis is a visualizer satisfying the RigidBodySim visualizer interface.

state is a RigidBodyDynamics.MechanismState,

representing the state of the mechanism that was simulated, and will be modified during the visualization.

sol is a DiffEqBase.ODESolution obtained from a RigidBodySim.jl simulation.

animate accepts the following keyword arguments:

max_fps: the maximum number of frames per second to draw. Default: 60.0.

realtime_rate: can be used to slow down or speed up playback compared to wall time. A realtime_rate of 2 will result in playback that is sped up 2x. Default: 1.0.

pause_pollint: how often to poll for commands coming from the director window when playback is paused. Default: 0.05.

Examples

Visualizing the result of a simulation of the passive dynamics of an Acrobot (double pendulum) at half speed:

using RigidBodySim, RigidBodyDynamics, MechanismGeometries
import RigidBodyTreeInspector: Visualizer
urdf = Pkg.dir("RigidBodySim", "test", "urdf", "Acrobot.urdf")
mechanism = parse_urdf(Float64, urdf)
state = MechanismState(mechanism)
set_configuration!(state, [0.1; 0.2])
problem = ODEProblem(Dynamics(mechanism), state, (0., 2.))
sol = solve(problem, Vern7())
vis = Visualizer(mechanism, visual_elements(mechanism, URDFVisuals(urdf)))
window(vis)
animate(vis, state, sol; realtime_rate = 0.5)

# output

source

Visualizers

RigidBodyTreeInspector

RigidBodySim provides an interface for the RigidBodyTreeInspector.jl visualizer, built on top of Director.

The RigidBodyTreeInspector visualizer handle is a DrakeVisualizer.Visualizer. The easiest way to create a Visualizer for a robot is from a URDF, which can be parsed by RigidBodyTreeInspector's parse_urdf function.

RigidBodySim.Visualization.VisualizerInterface.window — Method.

window(vis; reuse)

Open a new director visualizer window.

The director instance will be started with a script that handles communication between RigidBodySim.jl and the director instance.

Supports the following keyword arguments:

reuse: skip opening a new visualizer window if one is already open.

source

MeshCatMechanisms

RigidBodySim has support for the MeshCatMechanisms.jl MechanismVisualizer. Support is currently non-interactive, i.e., the simulation cannot be paused or terminated through the visualizer.

Visualizer interface

RigidBodySim visualizers must implement the following interface:

RigidBodySim.Visualization.VisualizerInterface.visualize

RigidBodySim.Visualization.VisualizerInterface.window

In addition, visualizers should also implement the following:

RigidBodySim.Visualization.VisualizerInterface.SimulationCommands — Type.

type SimulationCommands

Stores visualizer-independent commands used to control the simulation.

A module providing a specific visualizer instance, say MyVisualizer, should provide a SimulationCommands constructor method with the signature

SimulationCommands(vis::MyVisualizer)

which returns a SimulationCommands object (to be constructed using SimulationCommands() in addition to performing any visualizer-dependent setup.

source

RigidBodySim.Visualization.VisualizerInterface.isinteractive — Function.

isinteractive(vis)

Return whether or not visualizer vis supports user interaction (e.g. sending pause and terminate commands).

source

Utilities

RigidBodySim.Core.configuration_renormalizer — Function.

configuration_renormalizer(state)
configuration_renormalizer(state, condition)

configuration_renormalizer can be used to create a callback that projects the configuration of a mechanism's state onto the configuration manifold. This may be necessary for mechanism's with e.g. quaternion-parameterized orientations as part of their joint configuration vectors, as numerical integration can cause the configuration to drift away from the unit norm constraints.

The callback is implemented as a DiffEqCallbacks.DiscreteCallback By default, it is called at every integrator time step.

source